MLA203 FINAL CODE/ai-future-trading/requirements.txt
flask
python-dotenv
python-binance==1.0.16
openai
pandas
matplotlib
numpy
requests
